♻️  NICE-16 events=>slug(client); useNotion(options)
