+++ sites/jeromefitzgerald.com/src/hooks/useNotion.ts
@@ -2,8 +2,12 @@ import useSWR from 'swr'
 
 const fetcher = (url) => fetch(url).then((res) => res.json())
 
-function useNotion(routeType) {
-  const { data, error, isLoading } = useSWR(`/api/v1/cms/${routeType}`, fetcher)
+function useNotion(routeType, options = {}) {
+  const { data, error, isLoading } = useSWR(
+    `/api/v1/cms/${routeType}`,
+    fetcher,
+    options
+  )
   return {
     data,
     isLoading,
